--- Sensor/app/src/main/res/layout/activity_main.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.leekitack.sensor.MainActivity">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="충돌 감지 중입니다."
-        android:id="@+id/notify"
-        android:textSize="20dp" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="대기 시간은 20초 입니다."
-        android:id="@+id/timer"
-        android:layout_below="@+id/notify"
-        android:layout_alignParentStart="true"
-        android:layout_marginTop="44dp"
-        android:textSize="20dp" />
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="음성 인식 결과입니다"
-        android:id="@+id/voice"
-        android:layout_below="@+id/timer"
-        android:layout_alignParentStart="true"
-        android:layout_marginTop="55dp"
-        android:textSize="20dp" />
-</RelativeLayout>
